Finalized thing ready to commit
